+++ src/Privat24BusinessApi.php
@@ -2,6 +2,7 @@
 
 namespace Sashalenz\Privat24BusinessApi;
 
+use Sashalenz\Privat24BusinessApi\ApiModels\Payments;
 use Sashalenz\Privat24BusinessApi\ApiModels\Statements;
 
 class Privat24BusinessApi
@@ -9,5 +10,10 @@ class Privat24BusinessApi
     public static function statements(): Statements
     {
         return new Statements;
+    }
+
+    public static function payments(): Payments
+    {
+        return new Payments;
     }
 }