+++ src/RequestData/Payments/GetPaymentRequest.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Sashalenz\Privat24BusinessApi\RequestData\Payments;
+
+use Spatie\LaravelData\Data;
+
+class GetPaymentRequest extends Data
+{
+    public function __construct(
+        public string $ref,
+    ) {}
+}